+++ Cyotek.Windows.Forms.ColorPicker.Demo/ScreenColorPickerDemoForm.cs
@@ -1,38 +1,67 @@
-﻿using System;
+// Cyotek Color Picker Controls Library
+// http://cyotek.com/blog/tag/colorpicker
+
+// Copyright (c) 2013-2021 Cyotek Ltd.
+
+// This work is licensed under the MIT License.
+// See LICENSE.TXT for the full text
+
+// Found this code useful?
+// https://www.cyotek.com/contribute
+
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Windows.Forms;
 
 namespace Cyotek.Windows.Forms.ColorPicker.Demo
 {
-  // Cyotek Color Picker controls library
-  // Copyright © 2013-2017 Cyotek Ltd.
-  // http://cyotek.com/blog/tag/colorpicker
-
-  // Licensed under the MIT License. See license.txt for the full text.
-
-  // If you use this code in your applications, donations or attribution are welcome
-
   internal partial class ScreenColorPickerDemoForm : BaseForm
   {
-    #region Constructors
+    #region Public Constructors
 
     public ScreenColorPickerDemoForm()
     {
       this.InitializeComponent();
     }
 
-    #endregion
+    #endregion Public Constructors
 
-    #region Methods
+    #region Private Methods
 
-    private void closeToolStripMenuItem_Click(object sender, EventArgs e)
+    private void CaptureButton_Click(object sender, EventArgs e)
+    {
+      screenColorPicker.CaptureMouse();
+    }
+
+    private void CloseToolStripMenuItem_Click(object sender, EventArgs e)
     {
       this.Close();
     }
 
-    private void screenColorPicker1_ColorChanged(object sender, EventArgs e)
+    private void ReleaseButton_Click(object sender, EventArgs e)
     {
-      optionsSplitContainer.Panel2.BackColor = screenColorPicker1.Color;
+      screenColorPicker.ReleaseMouse();
     }
 
-    #endregion
+    private void ScreenColorPicker_ColorChanged(object sender, EventArgs e)
+    {
+      colorPreviewBox.Color = screenColorPicker.Color;
+    }
+
+    private void ScreenColorPicker_Selected(object sender, EventArgs e)
+    {
+      eventsListBox.AddEvent((Control)sender, nameof(ScreenColorPicker.Selected));
+    }
+
+    private void ScreenColorPicker_Selecting(object sender, CancelEventArgs e)
+    {
+      eventsListBox.AddEvent((Control)sender, nameof(ScreenColorPicker.Selecting), new Dictionary<string, object>
+      {
+        { nameof(CancelEventArgs.Cancel) ,e.Cancel }
+      });
+    }
+
+    #endregion Private Methods
   }
 }